Update engine improvements
common_service.sh is also run in parallel to update_engine.py. This
way it is easy to communicate between the two processes and removes
many overheads such as calling common_service.sh over and over
again which slowed things down.

common_service.sh executes operations directly by spying on dbus
and sends output to update_engine.py which later sends them to the
client. This seems to be an overhead but currently there's no way
to reduce this without converting the entire project into Python.

UpdateAttempter is further improved to support check_for_update,
but it directly upgrades the OS if an update is available (BUG).

Readme is changed to add usage info.